Automated Pollution Sensor
One of the major causes of air pollution is the harmful smoke produced by vehicles in India.

Thus, the idea of creating car pollution checkers using 'Arduino' and 'MQ-2 sensors' came up.
The main aim is to construct a cost-efficient mechanism to help combat this issue.
This sensor will detect LPG, butane, propane, methane, alcohol, and hydrogen- gases which are commonly present in the engine smoke.
The device will then notify the owner if the concentration of the above-mentioned gases is above the fixed level. The owner can then take action as needed.
One of the chief advantages of this device is that it is inexpensive and very efficient in its purpose.
Key Features-
1. Gas Level Monitoring
2. Visual Alerts (Green LED, Red LED)
3. Auditory Alerts (Buzzer)
4. Serial Output
This system can also be utilised in homes, laboratories, or industrial settings to detect and alert users about potentially hazardous gas leakage, ensuring safety and early warning.